Fix FacetTypeExtractingCamelProcessor
